Alpha release

+++ src/main/java/cogent/books/security/services/UserDetailsImpl.java
@@ -45,7 +45,7 @@ public class UserDetailsImpl implements UserDetails {
 		List<GrantedAuthority> authorities = user.getRoles().stream()
 				.map(role -> new SimpleGrantedAuthority(role.getName().name())).collect(Collectors.toList());
 
-		return new UserDetailsImpl(user.getId(), user.getUsername(), user.getEmail(), user.getPassword(), user.getName(), authorities);
+		return new UserDetailsImpl((long) user.getId(), user.getUsername(), user.getEmail(), user.getPassword(), user.getName(), authorities);
 	}
 
 	@Override
